Free allocated resources when doing a reset and on closing
Make an effort to fully free allocated resources when doing a reset
and when closing the device (_stop function).

Three property-related variables, use_buffer_pool, use_hardware_overlay
and vsync are now defined seperately from the property variables
because their value may be changed during hardware configuration. When
the element start function is called or when it is reset the property
values are copied into the seperately defined variables.